Issue #4437: Modified ImportOrderCheckTest.java and moved its input files to the importorder subdirectory
