--- src/test/resources/com/puppycrawl/tools/checkstyle/checks/imports/InputImportOrder_NoFailureForRedundantImports.java
@@ -1,5 +1,0 @@
-package com.puppycrawl.tools.checkstyle.checks.imports;
-
-public class InputImportOrder_NoFailureForRedundantImports {
-}
-
