+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/imports/importorder/InputImportOrderCaseInsensitive.java
@@ -1,4 +1,4 @@
-package com.puppycrawl.tools.checkstyle.checks.imports;
+package com.puppycrawl.tools.checkstyle.checks.imports.importorder;
 
 import java.io.File;
 import java.io.InputStream;
