+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/imports/importorder/InputImportOrder_Above.java
@@ -1,20 +1,20 @@
-package com.puppycrawl.tools.checkstyle.checks.imports;
+package com.puppycrawl.tools.checkstyle.checks.imports.importorder;
 
+import static java.awt.Button.ABORT;
+import static javax.swing.WindowConstants.*;
+import static java.awt.Button.ABORT;
 import java.awt.Button;
 import java.awt.Frame;
 import java.awt.Dialog;
 import java.awt.event.ActionEvent;
 import javax.swing.JComponent;
 import javax.swing.JTable;
-import static java.awt.Button.ABORT;
-import static javax.swing.WindowConstants.*;
-import static java.awt.Button.ABORT;
 
-import static java.io.File.createTempFile;
 import java.io.File;
+import static java.io.File.createTempFile;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.Reader;
 
-public class InputImportOrder_Under {
+public class InputImportOrder_Above {
 }
