+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/imports/importorder/InputImportOrder_Bottom.java
@@ -1,14 +1,10 @@
-package com.puppycrawl.tools.checkstyle.checks.imports;
-
-import static java.io.File.createTempFile;
-import static java.awt.Button.ABORT;
-import static javax.swing.WindowConstants.*;
+package com.puppycrawl.tools.checkstyle.checks.imports.importorder;
 
 import java.awt.Button;
 import java.awt.Dialog;
 import java.awt.Frame;
 import java.awt.event.ActionEvent;
-/***comment test***/
+
 import java.io.IOException;
 import java.io.InputStream;
 
@@ -17,7 +13,12 @@ import javax.swing.JTable;
 
 import static java.io.File.*;
 import java.io.File;
+
+import static java.io.File.createTempFile;
+import static java.awt.Button.ABORT;
+import static javax.swing.WindowConstants.*;
+
 import java.io.Reader;
 
-public class InputImportOrder_Top {
+public class InputImportOrder_Bottom {
 }
