+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/imports/importorder/InputImportOrder_InFlow.java
@@ -1,4 +1,4 @@
-package com.puppycrawl.tools.checkstyle.checks.imports;
+package com.puppycrawl.tools.checkstyle.checks.imports.importorder;
 
 import java.awt.Button;
 import static java.awt.Button.ABORT;
