+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/imports/importorder/InputImportOrder_NoFailureForRedundantImports.java
@@ -1,0 +1,5 @@
+package com.puppycrawl.tools.checkstyle.checks.imports.importorder;
+
+public class InputImportOrder_NoFailureForRedundantImports {
+}
+
