+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/imports/importorder/InputImportOrder_Under.java
@@ -1,0 +1,20 @@
+package com.puppycrawl.tools.checkstyle.checks.imports.importorder;
+
+import java.awt.Button;
+import java.awt.Frame;
+import java.awt.Dialog;
+import java.awt.event.ActionEvent;
+import javax.swing.JComponent;
+import javax.swing.JTable;
+import static java.awt.Button.ABORT;
+import static javax.swing.WindowConstants.*;
+import static java.awt.Button.ABORT;
+
+import static java.io.File.createTempFile;
+import java.io.File;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.Reader;
+
+public class InputImportOrder_Under {
+}
