+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/imports/importorder/InputImportOrder_Wildcard.java
@@ -1,9 +1,9 @@
-package com.puppycrawl.tools.checkstyle.checks.imports;
+package com.puppycrawl.tools.checkstyle.checks.imports.importorder;
 // groups are configured as follows
 // com.puppycrawl,*,java
 // the trailing javax.crypto.Cipher; should be flagged as an error.
 
-import com.puppycrawl.tools.checkstyle.checks.imports.InputImportOrder_Above;
+import com.puppycrawl.tools.checkstyle.checks.imports.importorder.InputImportOrder_Above;
 import javax.crypto.BadPaddingException;
 import java.util.List; //comment test
 import javax.crypto.Cipher;
